Upravy v tychto suboroch predchadzaju predchadzajucemu commitu, len som ich zabudol commitnut.
Dokoncena implementacia ziskavania vyskovych SRTM dat zo USGS. Vytvorena implementacia mapovej reprezentacie ktora je zavisla na vyskovych datach a vklada ich do atributov vrcholov. Zaroven dorobena impelemntacie uzivatelskeho modelu, aby vyskove data z atributov vedel vyuzit. Na zahrnutie prevysenie prekonaneho medzi dvomi vrcholmi vuziva analizu tempa pre bezcov do kopca od Anthony Kay-a. Vysledna funkcia tejto analyzy vracia tempo elitneho bezca pre konretnu hodnotu prevysenia.

Dalej sa v commitu nachadzaju aj upravy dalsich tried na ktore pocas implementacie vznikli ako riesenia problemov, ci logickych zaskobrtnuti minuleho mateja.

+++ Program/Optepafi/Optepafi/Models/TemplateMan/Templates/Orienteering_ISOM_2017_2.cs
@@ -56,10 +56,10 @@ public class Orienteering_ISOM_2017_2 : ITemplate<Orienteering_ISOM_2017_2.Verte
     // 3 bits
     public enum ManMadeLinearObstacles {Trench_215 = 1, Wall_513, ImpassableWall_515, Fence_516, ImpassableFence_518, ProminentLineFeature_528, ImpassableProminentLineFeature_529}
 
-    public struct VertexAttributes(MapCoordinates position, short elevation = 0) : IVertexAttributes
+    public struct VertexAttributes(MapCoordinates position, double elevation = 0) : IVertexAttributes
     {
         public readonly MapCoordinates Position = position;
-        public readonly short Elevation = elevation;
+        public readonly double Elevation = elevation;
     }
 
     public struct EdgeAttributes((Grounds? ground, Boulders? boulders, Stones? stones, Water? water, VegetationAndManMade? vegetationAndManMade, VegetationGoodVis? vegetationGoodVis) leftSurrounding,
